configure: error if readline not found, but requested explicitly
If --enable-readline option is passed to configure, but no readline
support is found, previous behavior was to silently continue
and not enable readline support. The new behavior is to error out
if readline was explicitly enabled, but there's no library that
provides it. Users are expected to either drop the --enable-readline
flag or install the missing dependency.