main: switch to WAL mode if custom WAL methods were used
This patch adds best-effort logic to switch to WAL journaling mode
as soon as a database connection is open, if the database was open
with custom WAL methods defined.
It's performed by calling an extra "pragma journal_mode=wal" query
at the end of the initialization process, and thus it's not guaranteed
to succeed, e.g. in readonly mode.