replication: handle NEED_SNAPSHOT error
If we get it, we must ask the primary for a snapshot file
instead of trying to fetch frames from a stream.

NOTICE: snapshots can be large, so we don't assume they fit
in memory. In sqld, they are kept in a temporary file stored
in the database directory. In embedded replicas, we currently
just use the temporary directory. If we decide that the db
directory is a better idea, we need to push the db path
further in the interface, and that can be done as follow-up.