Run apt update on sqld runtime image
This is a best effort attempt to run sqld with the same libraries that
it was built with. If anyone bases their images on this one and runs
apt update, and then install dependencies (e.g. because they need it
for extensions), then they risk running sqld with different dependencies
then it was built with.

Publishing the image with apt index updated allow images that depend
on this one to install further dependencies with the same version sqld
was built with. Of course apt update from the chef layer may get out
of sync with the runtime one, but this is a problem that we can solve,
since we control the life cycle of the image.